+++ partner/src/epilot/models/operations/resendpartnerinvitation.py
@@ -3,13 +3,13 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import partner as shared_partner
+from ...models.shared import partner as shared_partner
 from dataclasses_json import Undefined, dataclass_json
 from enum import Enum
 from epilot import utils
 from typing import Optional
 
-class ResendPartnerInvitationRequestBodyLanguage(str, Enum):
+class Language(str, Enum):
     r"""Language for partner invitation email"""
     EN = 'en'
     DE = 'de'
@@ -18,7 +18,7 @@ class ResendPartnerInvitationRequestBodyLanguage(str, Enum):
 @dataclass_json(undefined=Undefined.EXCLUDE)
 @dataclasses.dataclass
 class ResendPartnerInvitationRequestBody:
-    language: Optional[ResendPartnerInvitationRequestBodyLanguage] = dataclasses.field(default=ResendPartnerInvitationRequestBodyLanguage.EN, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('language'), 'exclude': lambda f: f is None }})
+    language: Optional[Language] = dataclasses.field(default=Language.EN, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('language'), 'exclude': lambda f: f is None }})
     r"""Language for partner invitation email"""