+++ partner/src/epilot/models/shared/partnerinvitationpayload.py
@@ -7,7 +7,7 @@ from enum import Enum
 from epilot import utils
 from typing import Optional
 
-class PartnerInvitationPayloadLanguage(str, Enum):
+class Language(str, Enum):
     r"""Language for partner invitation email"""
     EN = 'en'
     DE = 'de'
@@ -16,7 +16,7 @@ class PartnerInvitationPayloadLanguage(str, Enum):
 @dataclass_json(undefined=Undefined.EXCLUDE)
 @dataclasses.dataclass
 class PartnerInvitationPayload:
-    language: Optional[PartnerInvitationPayloadLanguage] = dataclasses.field(default=PartnerInvitationPayloadLanguage.EN, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('language'), 'exclude': lambda f: f is None }})
+    language: Optional[Language] = dataclasses.field(default=Language.EN, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('language'), 'exclude': lambda f: f is None }})
     r"""Language for partner invitation email"""